# docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_HongJungWan_recruit-process-engine-back_internal_applicant_dto_request.BulkUpdateStageRequest:
    properties:
      ids:
        items:
          type: integer
        type: array
      reason:
        type: string
      stage:
        type: string
    required:
    - ids
    - stage
    type: object
  github_com_HongJungWan_recruit-process-engine-back_internal_applicant_dto_request.UpdateStageRequest:
    properties:
      reason:
        type: string
      stage:
        type: string
    required:
    - stage
    type: object
  github_com_HongJungWan_recruit-process-engine-back_internal_applicant_dto_response.ApplicantDetail:
    properties:
      application_id:
        type: integer
      created_at:
        type: string
      created_by:
        type: string
      current_stage:
        type: string
      education:
        type: string
      email:
        type: string
      experience:
        type: string
      name:
        type: string
      phone:
        type: string
      tech_stack:
        type: string
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
  github_com_HongJungWan_recruit-process-engine-back_internal_applicant_dto_response.ApplicantItem:
    properties:
      application_id:
        type: integer
      current_stage:
        type: string
      email:
        type: string
      name:
        type: string
    type: object
  github_com_HongJungWan_recruit-process-engine-back_internal_applicant_dto_response.BulkUpdateResponse:
    properties:
      updated:
        type: integer
    type: object
  github_com_HongJungWan_recruit-process-engine-back_internal_applicant_dto_response.ListApplicantsResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/github_com_HongJungWan_recruit-process-engine-back_internal_applicant_dto_response.ApplicantItem'
        type: array
      page:
        type: integer
      size:
        type: integer
      total:
        type: integer
    type: object
  github_com_HongJungWan_recruit-process-engine-back_internal_applicant_dto_response.StageHistoryItem:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      history_id:
        type: integer
      stage:
        type: string
      status:
        type: string
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
  github_com_HongJungWan_recruit-process-engine-back_internal_applicant_dto_response.UpdateStageResponse:
    properties:
      application_id:
        type: integer
      new_stage:
        type: string
      old_stage:
        type: string
      updated_at:
        type: string
    type: object
  github_com_HongJungWan_recruit-process-engine-back_internal_offer_dto_request.CreateApprovalsRequest:
    properties:
      approver_ids:
        items:
          type: integer
        type: array
    required:
    - approver_ids
    type: object
...
  github_com_HongJungWan_recruit-process-engine-back_internal_offer_dto_request.ProcessApprovalRequest:
    properties:
      comment:
        type: string
      status:
        type: string
    required:
    - status
    type: object
  github_com_HongJungWan_recruit-process-engine-back_internal_offer_dto_response.ApprovalHistoryItem:
    properties:
      approval_id:
        type: integer
      approver_id:
        type: integer
      comment:
        type: string
      decided_at:
        type: string
      requested_at:
        type: string
      status:
        type: string
    type: object
...
  github_com_HongJungWan_recruit-process-engine-back_internal_offer_dto_response.ProcessApprovalResponse:
    properties:
      approval_id:
        type: integer
      decided_at:
        type: string
      status:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: test@example.com
    name: 홍정완
  description: 채용 프로세스 엔진 백엔드 API 문서
  title: Recruit Process Engine API
  version: "1.0"
paths:
  /applicants:
    get:
      description: 지원자 목록과 총 개수를 조회한다.
      parameters:
      - description: 페이지 번호
        in: query
        name: page
        required: true
        type: integer
      - description: 페이지당 항목 수
        in: query
        name: size
        required: true
        type: integer
      - description: 단계 필터 (서류 접수, 기술 면접…)
        in: query
        name: stage
        type: string
      - description: 이름 또는 이메일 키워드 검색
        in: query
        name: keyword
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_HongJungWan_recruit-process-engine-back_internal_applicant_dto_response.ListApplicantsResponse'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: 지원자 목록 조회
      tags:
      - Applicants
  /applicants/{application_id}:
    get:
      description: 단일 지원자 정보를 조회한다.
      parameters:
      - description: 지원자 식별자
        in: path
        name: application_id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_HongJungWan_recruit-process-engine-back_internal_applicant_dto_response.ApplicantDetail'
        "404":
          description: Not Found
      summary: 지원자 상세 조회
      tags:
      - Applicants
  /applicants/{application_id}/history:
    get:
      description: 해당 지원자의 전형 단계 변경 이력을 조회한다.
      parameters:
      - description: 지원자 식별자
        in: path
        name: application_id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_HongJungWan_recruit-process-engine-back_internal_applicant_dto_response.StageHistoryItem'
            type: array
        "500":
          description: Internal Server Error
      summary: 지원자 단계 변경 이력 조회
      tags:
      - Applicants
  /applicants/{application_id}/stage:
    patch:
      description: 지원자 전형 단계를 수정한다.
      parameters:
      - description: 지원자 식별자
        in: path
        name: application_id
        required: true
        type: integer
      - description: 업데이트할 단계 정보
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_HongJungWan_recruit-process-engine-back_internal_applicant_dto_request.UpdateStageRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_HongJungWan_recruit-process-engine-back_internal_applicant_dto_response.UpdateStageResponse'
        "400":
          description: Bad Request
      summary: 지원자 단계 업데이트
      tags:
      - Applicants
  /applicants/stages/bulk-update:
    post:
      description: 여러 지원자의 전형 단계를 일괄 수정한다.
      parameters:
      - description: 일괄 업데이트 요청
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_HongJungWan_recruit-process-engine-back_internal_applicant_dto_request.BulkUpdateStageRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_HongJungWan_recruit-process-engine-back_internal_applicant_dto_response.BulkUpdateResponse'
        "400":
          description: Bad Request
      summary: 지원자 단계 일괄 업데이트
      tags:
      - Applicants
...
  /offers/{offer_id}/approvals:
    get:
      description: 오퍼의 승인(결재) 이력 목록을 반환합니다.
      parameters:
      - description: 오퍼 식별자
        in: path
        name: offer_id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_HongJungWan_recruit-process-engine-back_internal_offer_dto_response.ApprovalHistoryItem'
            type: array
        "500":
          description: Internal Server Error
      summary: 오퍼별 승인 이력 조회
      tags:
      - Approvals
    post:
      description: 오퍼 승인(결재) 요청을 생성한다.
      parameters:
      - description: 오퍼 식별자
        in: path
        name: offer_id
        required: true
        type: integer
      - description: 승인자 ID 리스트
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_HongJungWan_recruit-process-engine-back_internal_offer_dto_request.CreateApprovalsRequest'
      responses:
        "200":
          description: OK
          schema:
            items:
              items:
                $ref: '#/definitions/github_com_HongJungWan_recruit-process-engine-back_internal_offer_dto_response.ApprovalHistoryItem'
              type: array
            type: array
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: 승인 요청 생성
      tags:
      - Approvals
  /offers/{offer_id}/approvals/{approval_id}:
    put:
      description: 오퍼 승인(결재) 요청을 처리한다.
      parameters:
      - description: 승인 요청 식별자
        in: path
        name: approval_id
        required: true
        type: integer
      - description: 승인 상태 및 코멘트
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_HongJungWan_recruit-process-engine-back_internal_offer_dto_request.ProcessApprovalRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_HongJungWan_recruit-process-engine-back_internal_offer_dto_response.ProcessApprovalResponse'
        "400":
          description: Bad Request
      summary: 승인 처리
      tags:
      - Approvals
...
swagger: "2.0"